Implement saving a Gig

+++ GigHub/ViewModels/GigFormViewModel.cs
@@ -1,15 +1,15 @@
 ﻿using GigHub.Models;
 using System.Collections.Generic;
-
+using System.ComponentModel.DataAnnotations;
 
 namespace GigHub.ViewModels
 {
     public class GigFormViewModel
-    {
+    {   
         public string Venue { get; set; }
         public string Date { get; set; }
         public string Time { get; set; }
-        public int Genre { get; set; }
+        public byte Genre { get; set; }
         public IEnumerable<Genre> Genres { get; set; }
     }
-}
+}
